fix AttributeError: module 'numpy' has no attribute 'int'

+++ pymatgen/util/coord_cython.pyx
@@ -251,7 +251,7 @@ def coord_list_mapping_pbc(subset, superset, atol=1e-8, pbc=(True, True, True)):
     Returns:
         list of indices such that superset[indices] = subset
     """
-    inds = np.zeros(len(subset), dtype=np.int) - 1
+    inds = np.zeros(len(subset)) - 1
     subset = np.atleast_2d(subset)
     superset = np.atleast_2d(superset)
